Stable version after massive refactoring. All functions are ready.
