+++ src/main/java/com/javarush/quest/ivanilov/config/BlackEarthQuest.java
@@ -25,7 +25,7 @@ public class BlackEarthQuest {
                     Break:Break;
                     
                     Event4:Шулер.;
-                    Q:Ты вернулся во двор после нескольких часов работы в поле. Руки-крюки, спину тянет. Зычный смех деда привлекает твоё внимание - на завалинке он с друзьями играет в преферанс. Ты вот уже 20 лет мечтаешь научиться в преферанс.; 
+                    Q:Ты вернулся во двор после нескольких часов работы в поле. Руки-крюки, спину тянет. Зычный смех деда привлекает твоё внимание - на завалинке он с друзьями играет в преферанс. Ты вот уже 20 лет мечтаешь научиться в преферанс.;
                     Ans:Мечты нужно реализовывать - иду к дедам.^Event6;
                     Ans:Азартные игры опасны, лучше завалюсь на печь.^Event7;
                     Break:Break;
