--- src/main/java/com/javarush/quest/ivanilov/constants/Targets.java
@@ -1,39 +1,0 @@
-package com.javarush.quest.ivanilov.constants;
-
-public class Targets {
-    public static final String MAIN_JSP = "main.jsp";
-    public static final String LOGIN_JSP = "login.jsp";
-    public static final String USERS_JSP = "users.jsp";
-    public static final String INDEX_JSP = "index.jsp";
-    public static final String SIGNUP_JSP = "signup.jsp";
-    public static final String ERROR_MESSAGE_JSP = "errorMessage.jsp";
-    public static final String USER_MODIFY_JSP = "userModify.jsp";
-    public static final String SUCCESS_MESSAGE_JSP = "successMessage.jsp";
-    public static final String USER_DELETE_JSP = "userDelete.jsp";
-    public static final String USER_CREATE_JSP = "userCreate.jsp";
-    public static final String MAIN = "/main";
-    public static final String ACCOUNT = "/account";
-    public static final String USER_CREATE = "/userCreate";
-    public static final String USER_DELETE = "/userDelete";
-    public static final String USER_DELETE_ETC = "/userDelete/*";
-    public static final String INDEX = "";
-    public static final String LOGIN = "/login";
-    public static final String LOGOUT = "/logout";
-    public static final String USER_MODIFY = "/userModify";
-    public static final String USER_MODIFY_ETC = "/userModify/*";
-    public static final String SIGNUP = "/signup";
-    public static final String USERS = "/users";
-    public static final String ACCOUNT_JSP = "account.jsp";
-    public static final String PLAY = "/play";
-    public static final String NEW_GAME = "/newGame";
-    public static final String SHOW_RESULT = "/showResult";
-    public static final String DISPATCH_TASK = "/dispatchTask";
-    public static final String QUESTION = "/question";
-    public static final String FIGHT = "/fight";
-    public static final String QUESTION_JSP = "question.jsp";
-    public static final String FIGHT_JSP = "fight.jsp";
-    public static final String SHOW_RESULT_JSP = "showResult.jsp";
-    public static final String CONSTRUCTOR_JSP = "constructor.jsp";
-    public static final String CONSTRUCTOR = "/constructor";
-    public static final String ALL = "/**";
-}
