+++ src/main/java/com/javarush/quest/ivanilov/controllers/EventService.java
@@ -1,4 +1,4 @@
-package com.javarush.quest.ivanilov.services;
+package com.javarush.quest.ivanilov.controllers;
 
 import com.javarush.quest.ivanilov.entities.game.Event;
 import com.javarush.quest.ivanilov.repositories.EventRepository;
