+++ src/main/java/com/javarush/quest/ivanilov/servlets/DeleteUserServlet.java
@@ -1,19 +1,20 @@
 package com.javarush.quest.ivanilov.servlets;
 
-import com.javarush.quest.ivanilov.constants.Attributes;
-import com.javarush.quest.ivanilov.constants.Messages;
-import com.javarush.quest.ivanilov.constants.Targets;
+import com.javarush.quest.ivanilov.utils.constants.Attributes;
+import com.javarush.quest.ivanilov.utils.constants.Jsp;
+import com.javarush.quest.ivanilov.utils.constants.Messages;
+import com.javarush.quest.ivanilov.utils.constants.Targets;
 import com.javarush.quest.ivanilov.entities.users.User;
-import com.javarush.quest.ivanilov.services.AuthorizationService;
-import com.javarush.quest.ivanilov.services.UserService;
+import com.javarush.quest.ivanilov.controllers.AuthorizationService;
+import com.javarush.quest.ivanilov.controllers.UserService;
 import com.javarush.quest.ivanilov.utils.Navigator;
-import com.javarush.quest.ivanilov.utils.RequestUtils;
-import jakarta.servlet.*;
-import jakarta.servlet.http.*;
-import jakarta.servlet.annotation.*;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
-import java.util.Map;
 
 @WebServlet(name = "DeleteUserServlet", value = Targets.USER_DELETE_ETC)
 public class DeleteUserServlet extends HttpServlet {
@@ -21,36 +22,32 @@ public class DeleteUserServlet extends HttpServlet {
     private final UserService userService = UserService.USER_SERVICE;
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        auth.authorizeAndProceed(req, resp, Targets.LOGIN_JSP);
         User user = (User) req.getSession().getAttribute(Attributes.USER);
 
         if (auth.isAdmin(user)) {
-            Map<String, String> attributes = RequestUtils.getAttributes(req.getQueryString());
-            long userId = Long.parseLong(attributes.getOrDefault(Attributes.USER_ID, Attributes.DEFAULT_ID));
+            long userId = Long.parseLong(req.getParameter(Attributes.USER_ID));
             User userToBeDeleted = userService.get(userId);
             req.setAttribute(Attributes.USER_TO_BE_MODIFIED, userToBeDeleted);
-            Navigator.dispatch(req, resp, Targets.USER_DELETE_JSP);
+            Navigator.dispatch(req, resp, Jsp.USER_DELETE_JSP);
         } else {
             req.setAttribute(Attributes.MESSAGE, new Messages().forbidden(user.getLogin()));
-            Navigator.dispatch(req, resp, Targets.ERROR_MESSAGE_JSP);
+            Navigator.dispatch(req, resp, Jsp.ERROR_MESSAGE_JSP);
         }
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        auth.authorizeAndProceed(req, resp, Targets.LOGIN_JSP);
         User user = (User) req.getSession().getAttribute(Attributes.USER);
 
         if (auth.isAdmin(user)) {
-            Map<String, String> attributes = RequestUtils.getAttributes(req);
-            long userId = Long.parseLong(attributes.getOrDefault(Attributes.USER_ID, Attributes.DEFAULT_ID));
+            long userId = Long.parseLong(req.getParameter(Attributes.USER_ID));
             User userToBeDeleted = userService.get(userId);
             userService.delete(userToBeDeleted);
             req.setAttribute(Attributes.MESSAGE, new Messages().userDeleted(userToBeDeleted));
-            Navigator.dispatch(req, resp, Targets.SUCCESS_MESSAGE_JSP);
+            Navigator.dispatch(req, resp, Jsp.SUCCESS_MESSAGE_JSP);
         } else {
             req.setAttribute(Attributes.MESSAGE, new Messages().forbidden(user.getLogin()));
-            Navigator.dispatch(req, resp, Targets.ERROR_MESSAGE_JSP);
+            Navigator.dispatch(req, resp, Jsp.ERROR_MESSAGE_JSP);
         }
     }
