+++ src/main/java/com/javarush/quest/ivanilov/servlets/MainServlet.java
@@ -1,30 +1,29 @@
 package com.javarush.quest.ivanilov.servlets;
 
-import com.javarush.quest.ivanilov.constants.Attributes;
-import com.javarush.quest.ivanilov.constants.Targets;
+import com.javarush.quest.ivanilov.controllers.GameService;
+import com.javarush.quest.ivanilov.controllers.QuestService;
 import com.javarush.quest.ivanilov.entities.game.Game;
 import com.javarush.quest.ivanilov.entities.game.Quest;
 import com.javarush.quest.ivanilov.entities.users.User;
-import com.javarush.quest.ivanilov.services.AuthorizationService;
-import com.javarush.quest.ivanilov.services.GameService;
-import com.javarush.quest.ivanilov.services.QuestService;
 import com.javarush.quest.ivanilov.utils.Navigator;
-import jakarta.servlet.*;
-import jakarta.servlet.http.*;
-import jakarta.servlet.annotation.*;
+import com.javarush.quest.ivanilov.utils.constants.Attributes;
+import com.javarush.quest.ivanilov.utils.constants.Jsp;
+import com.javarush.quest.ivanilov.utils.constants.Targets;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 import java.util.Collection;
 
 @WebServlet(name = "MainServlet", value = Targets.MAIN)
 public class MainServlet extends HttpServlet {
-
-    AuthorizationService auth = AuthorizationService.AUTHORIZATION_SERVICE;
     QuestService questService = QuestService.QUEST_SERVICE;
     GameService gameService = GameService.GAME_SERVICE;
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        auth.authorizeAndProceed(req, resp, Targets.LOGIN);
         Collection<Quest> quests = questService.getAll();
         req.setAttribute(Attributes.QUESTS, quests);
         User user = (User) req.getSession().getAttribute(Attributes.USER);
@@ -39,10 +38,6 @@ public class MainServlet extends HttpServlet {
         }
 
         req.setAttribute(Attributes.IS_IN_GAME, isInGame);
-        Navigator.dispatch(req, resp, Targets.MAIN_JSP);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        Navigator.dispatch(req, resp, Jsp.MAIN_JSP);
     }
 }
