+++ src/main/java/com/javarush/quest/ivanilov/utils/constants/Attributes.java
@@ -1,4 +1,4 @@
-package com.javarush.quest.ivanilov.constants;
+package com.javarush.quest.ivanilov.utils.constants;
 
 public class Attributes {
     public static final String LOGIN = "login";
@@ -10,16 +10,12 @@ public class Attributes {
     public static final String ROLES = "roles";
     public static final String USER_ID = "userId";
     public static final String MESSAGE = "message";
-    public static final String ROLE = "role";
-    public static final String DEFAULT_ID = "-1";
     public static final String QUESTS = "quests";
     public static final String QUEST_ID = "questId";
     public static final String CURRENT_GAME_NAME = "currentGameName";
     public static final String IS_IN_GAME = "isInGame";
     public static final String CURR_EVENT = "currEvent";
     public static final String OPTIONS = "options";
-    public static final String OPTION = "option";
-    public static final String EMPTY = "";
     public static final String OPTION_ID = "optionId";
     public static final String FIGHT = "fight";
     public static final String HIT_OPTION = "hitOption";
@@ -28,5 +24,6 @@ public class Attributes {
     public static final String GAMES_WON = "gamesWon";
     public static final String NUMBER_OF_GAMES_PLAYED = "numberOfGamesPlayed";
     public static final String QUEST_SAMPLE = "questSample";
-    public static final String CODE = "code";
+    public static final String QUEST_SCENARIO = "questScenario";
+    public static final String EVENT = "Event";
 }
