+++ src/main/java/com/javarush/quest/ivanilov/utils/constants/Jsp.java
@@ -1,0 +1,19 @@
+package com.javarush.quest.ivanilov.utils.constants;
+
+public class Jsp {
+    public static final String MAIN_JSP = "main.jsp";
+    public static final String LOGIN_JSP = "login.jsp";
+    public static final String USERS_JSP = "users.jsp";
+    public static final String INDEX_JSP = "index.jsp";
+    public static final String SIGNUP_JSP = "signup.jsp";
+    public static final String ERROR_MESSAGE_JSP = "errorMessage.jsp";
+    public static final String USER_MODIFY_JSP = "userModify.jsp";
+    public static final String SUCCESS_MESSAGE_JSP = "successMessage.jsp";
+    public static final String USER_DELETE_JSP = "userDelete.jsp";
+    public static final String USER_CREATE_JSP = "userCreate.jsp";
+    public static final String ACCOUNT_JSP = "account.jsp";
+    public static final String QUESTION_JSP = "question.jsp";
+    public static final String FIGHT_JSP = "fight.jsp";
+    public static final String SHOW_RESULT_JSP = "showResult.jsp";
+    public static final String CONSTRUCTOR_JSP = "constructor.jsp";
+}
