+++ src/main/java/com/javarush/quest/ivanilov/utils/constants/Logs.java
@@ -1,4 +1,4 @@
-package com.javarush.quest.ivanilov.constants;
+package com.javarush.quest.ivanilov.utils.constants;
 
 public class Logs {
     public static final String AUTHORIZED_SUCCESSFULLY = "%s has been authorized successfully.";
