+++ src/main/java/com/javarush/quest/ivanilov/utils/constants/Targets.java
@@ -1,0 +1,21 @@
+package com.javarush.quest.ivanilov.utils.constants;
+
+public class Targets {
+    public static final String MAIN = "/main";
+    public static final String ACCOUNT = "/account";
+    public static final String USER_CREATE = "/userCreate";
+    public static final String USER_DELETE_ETC = "/userDelete/*";
+    public static final String INDEX = "";
+    public static final String LOGIN = "/login";
+    public static final String LOGOUT = "/logout";
+    public static final String USER_MODIFY_ETC = "/userModify/*";
+    public static final String SIGNUP = "/signup";
+    public static final String USERS = "/users";
+    public static final String PLAY = "/play";
+    public static final String NEW_GAME = "/newGame";
+    public static final String SHOW_RESULT = "/showResult";
+    public static final String DISPATCH_TASK = "/dispatchTask";
+    public static final String QUESTION = "/question";
+    public static final String FIGHT = "/fight";
+    public static final String CONSTRUCTOR = "/constructor";
+}
