+++ src/main/webapp/constructor.jsp
@@ -4,7 +4,6 @@
 <html>
 <head>
     <title>Конструктор квестов</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
     <link rel="stylesheet" href="styles/styles.css">
 </head>
 <body class="typewriter">
