+++ src/main/webapp/login.jsp
@@ -12,9 +12,9 @@
     <fieldset>
         <legend>Пожалуйста, авторизуйтесь</legend>
         <label for="text">Логин</label>
-        <input type="text" id="text" name="login" placeholder="login" />
+        <input type="text" id="text" name="login" placeholder="login" required/>
         <label for="stacked-password">Password</label>
-        <input type="password" id="stacked-password" name="password" />
+        <input type="password" id="stacked-password" name="password" required/>
         <button type="submit" class="button button-blue">Авторизоваться</button>
     </fieldset>
 </form>
